[CLIENT-PY] Renamed TestContainer to IoTDBContainer.
